requirements.txt
python-dotenv==0.21.0
requests==2.28.2
youtube_dl==2021.12.17  # youtube
